dev-requirements.txt
protobuf==3.20.*
dagshub==0.3.42
hydra-core==1.2.0
hyperopt==0.2.7
joblib==1.1.0
mlflow==1.26.0
numpy==1.22.4
pandas==1.4.2
pandera==0.11.0
patsy==0.5.2
pytest-steps==1.8.0
scikit_learn==1.1.1
streamlit==1.9.0
xgboost==1.6.1
dvc==2.10.2
pydantic #==1.9.1
pytest==7.1.2
#bentoml==1.0.0a7
deepchecks
fsspec==2022.11.0
dvc[s3]
uvicorn
fastapi
altair<5